+++ pages/dashboards/tasks/index.tsx
@@ -21,7 +21,7 @@ import TotalReviews from "@/content/Dashboards/Tasks/TotalReviews";
 import ReviewGrowth from "@/content/Dashboards/Tasks/ReviewGrowth";
 import StarRatingBreakDown from "@/content/Dashboards/Tasks/StarRatingBreakDown";
 import ReviewSourceBreakDown from "@/content/Dashboards/Tasks/ReviewSourceBreakDown";
-import ReviewsTable from "@/content/Dashboards/Tasks/ReviewsTable";
+import ReviewsTable from "@/content/Dashboards/Tasks/ReviewsTable/ReviewsTable";
 import SourceGraph from "@/content/Dashboards/Tasks/SourceGraph";
 import SourceTable from "@/content/Dashboards/Tasks/SourceTable";
